added search button to search filters
